start of refactoring for AML
