+++ objectherkenning_openbare_ruimte/create_aml_environment/oor-environment.Dockerfile
@@ -1,0 +1,33 @@
+FROM mcr.microsoft.com/azureml/openmpi4.1.0-cuda11.8-cudnn8-ubuntu22.04 AS base-image
+
+# Upgrade and install system libraries
+RUN apt-get -y update \
+    && ACCEPT_EULA=Y apt-get upgrade -qq \
+    && apt-get -y install \
+        build-essential \
+        curl \
+    && rm -rf /var/lib/apt/lists/*
+
+RUN apt-get update
+RUN apt-get install ffmpeg libsm6 libxext6 libpq-dev -y
+
+
+WORKDIR /opt/app
+
+RUN conda create -n env python=3.8
+RUN echo "source activate env" > ~/.bashrc
+ENV PATH="/opt/miniconda/envs/env/bin:$PATH"
+
+RUN curl -sSL https://install.python-poetry.org | python3 -
+ENV PATH="/root/.local/bin:$PATH"
+RUN poetry config virtualenvs.create false
+
+COPY pyproject.toml .
+COPY poetry.lock .
+
+# Initialize Conda, activate environment and install poetry packages
+RUN /opt/miniconda/bin/conda init bash && \
+    . /opt/miniconda/etc/profile.d/conda.sh && \
+    conda activate env && \
+    poetry update --no-ansi --no-interaction && \
+    poetry install --no-ansi --no-interaction --no-root
